Keep only the node implementation
