+++ packages/expo-file-system/src/next/FileSystem.ts
@@ -1,6 +1,6 @@
 import ExpoFileSystem from './ExpoFileSystem';
 import { URI } from './FileSystem.types';
-import { dirname, extname, join } from './pathNode';
+import { dirname, extname, join } from './path';
 
 export class File extends ExpoFileSystem.FileSystemFile {
   constructor(uri: URI) {
